Update GitHub workflows and remove unnecessary packages
This commit modifies the fail-fast strategy in the static analysis workflow and removes the installation of 'predis/predis' and 'illuminate/contracts' packages across multiple workflow files. This results in a cleaner setup, less dependency on external packages, and a more effective fail-fast configuration. Now, the workflows will use a matrix strategy for the PHP version, targeting specifically PHP